refactor: move UserMenu component to its own file for better organization
